experiments/03-vulnerable-dependencies/secure-requirements.txt
# Secure Dependencies - Updated versions with security patches
# ✅ Production-ready versions with no known HIGH/CRITICAL CVEs

# Python runtime (latest stable versions)
Flask==3.0.0              # ✅ No known CVEs, includes security fixes
Werkzeug==3.0.1           # ✅ Patched path traversal vulnerabilities
Jinja2==3.1.3             # ✅ XSS vulnerabilities fixed
itsdangerous==2.1.2       # ✅ Security improvements

# Database
psycopg2-binary==2.9.9    # ✅ Latest stable, security patches applied

# HTTP client
requests==2.31.0          # ✅ SSRF and security vulnerabilities fixed
urllib3==2.1.0            # ✅ Latest stable with security patches
certifi==2023.11.17       # ✅ Updated certificate bundle

# Flask extensions
flask-cors==4.0.0         # ✅ Latest version

# Monitoring
prometheus-client==0.19.0 # ✅ Latest version

# Production server
gunicorn==21.2.0          # ✅ Security fixes applied

# Security improvements in these versions:
# ✅ All CRITICAL and HIGH CVEs patched
# ✅ Improved input validation
# ✅ Better security defaults
# ✅ Regular security audits
# ✅ Active maintenance and support
#
# Recommended practices:
# - Run `pip install --upgrade` regularly
# - Use tools like Safety, Trivy, Snyk for scanning
# - Subscribe to security advisories
# - Automate dependency updates with Dependabot
